Make FinalizableNode internal
