+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTAnyTypeDeclaration.java
@@ -35,9 +35,9 @@ public interface ASTAnyTypeDeclaration
             TypeParamOwnerNode,
             ASTBodyDeclaration,
             ASTTopLevelDeclaration,
-            FinalizableNode,
+            InternalInterfaces.FinalizableNode,
             JavadocCommentOwner,
-            NonLocalDeclarationNode {
+            InternalInterfaces.NonLocalDeclarationNode {
 
     @Override
     @NonNull
