+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTMethodOrConstructorDeclaration.java
@@ -32,7 +32,7 @@ public interface ASTMethodOrConstructorDeclaration
             TypeParamOwnerNode,
             GenericNode<JavaNode>,
             JavadocCommentOwner,
-            NonLocalDeclarationNode {
+            InternalInterfaces.NonLocalDeclarationNode {
 
 
     @Override
